requirements.txt
Django>=2
django-extensions>=2
djangorestframework==3.7.7
django-crispy-forms==1.7.0
jupyter
pandas
numpy
sqlalchemy
datetime
psycopg2
django_pandas
